refactor(web): isolate receipt simulation
